Move endpoint + move ComputeValidatorPerformance
