+++ beacon-chain/rpc/prysm/v1alpha1/beacon/BUILD.bazel
@@ -38,6 +38,7 @@ go_library(
         "//beacon-chain/operations/attestations:go_default_library",
         "//beacon-chain/operations/slashings:go_default_library",
         "//beacon-chain/p2p:go_default_library",
+        "//beacon-chain/rpc:go_default_library",
         "//beacon-chain/state:go_default_library",
         "//beacon-chain/state/stategen:go_default_library",
         "//beacon-chain/sync:go_default_library",
