+++ beacon-chain/rpc/prysm/validator/validator_performance_test.go
@@ -1,4 +1,4 @@
-package beacon
+package validator
 
 import (
 	"bytes"
@@ -14,6 +14,7 @@ import (
 	mock "github.com/prysmaticlabs/prysm/v4/beacon-chain/blockchain/testing"
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/core/epoch/precompute"
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/beacon"
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/state"
 	mockSync "github.com/prysmaticlabs/prysm/v4/beacon-chain/sync/initial-sync/testing"
 	"github.com/prysmaticlabs/prysm/v4/config/params"
@@ -27,7 +28,7 @@ import (
 
 func TestServer_GetValidatorPerformance(t *testing.T) {
 	t.Run("Syncing", func(t *testing.T) {
-		bs := &Server{
+		bs := &beacon.Server{
 			SyncChecker: &mockSync.Sync{IsSyncing: true},
 		}
 
@@ -56,7 +57,7 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 		require.NoError(t, headState.SetBalances([]uint64{100, 101, 102}))
 
 		offset := int64(headState.Slot().Mul(params.BeaconConfig().SecondsPerSlot))
-		bs := &Server{
+		bs := &beacon.Server{
 			HeadFetcher: &mock.ChainService{
 				State: headState,
 			},
@@ -110,7 +111,7 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 		headState = setHeadState(t, headState, publicKeys)
 
 		offset := int64(headState.Slot().Mul(params.BeaconConfig().SecondsPerSlot))
-		bs := &Server{
+		bs := &beacon.Server{
 			HeadFetcher: &mock.ChainService{
 				// 10 epochs into the future.
 				State: headState,
@@ -173,7 +174,7 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 		headState = setHeadState(t, headState, publicKeys)
 
 		offset := int64(headState.Slot().Mul(params.BeaconConfig().SecondsPerSlot))
-		bs := &Server{
+		bs := &beacon.Server{
 			HeadFetcher: &mock.ChainService{
 				// 10 epochs into the future.
 				State: headState,
@@ -242,7 +243,7 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 		require.NoError(t, headState.SetInactivityScores([]uint64{0, 0, 0}))
 		require.NoError(t, headState.SetBalances([]uint64{100, 101, 102}))
 		offset := int64(headState.Slot().Mul(params.BeaconConfig().SecondsPerSlot))
-		bs := &Server{
+		bs := &beacon.Server{
 			HeadFetcher: &mock.ChainService{
 				State: headState,
 			},
@@ -302,7 +303,7 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 		require.NoError(t, headState.SetInactivityScores([]uint64{0, 0, 0}))
 		require.NoError(t, headState.SetBalances([]uint64{100, 101, 102}))
 		offset := int64(headState.Slot().Mul(params.BeaconConfig().SecondsPerSlot))
-		bs := &Server{
+		bs := &beacon.Server{
 			HeadFetcher: &mock.ChainService{
 				State: headState,
 			},
@@ -362,7 +363,7 @@ func TestServer_GetValidatorPerformance(t *testing.T) {
 		require.NoError(t, headState.SetInactivityScores([]uint64{0, 0, 0}))
 		require.NoError(t, headState.SetBalances([]uint64{100, 101, 102}))
 		offset := int64(headState.Slot().Mul(params.BeaconConfig().SecondsPerSlot))
-		bs := &Server{
+		bs := &beacon.Server{
 			HeadFetcher: &mock.ChainService{
 				State: headState,
 			},
